refactor: extract functions to improve readability
